refactor: Add team_id property to Member interface
This commit adds the `team_id` property to the `Member` interface in the `types.tsx` file. The `team_id` property represents the ID of the team to which the member belongs. This change enhances the data structure and allows for better organization and management of team-related information in the application.