refactor: Add notifications functionality and improve user experience
This commit adds notifications functionality to the application, including actions for requesting to join a team, requesting to join a tournament, challenging a match, inviting to a tournament, inviting to a team, and sending informational notifications. It also includes the addition of notifications to the notification aside component. Additionally, it implements features such as kicking a player, banning a player, changing the coach, leaving a team, and showing hidden team and member details. The code changes also involve updating the member collection when the team name is changed and modifying the isAlreadyInTeam function in the authSlice to use the __name__ way, similar to the teamSlice. The commit also includes updates to the Firebase rules for teams, notifications, and users to ensure proper access control.