refactor: Update Notification interface to include action property
This commit updates the Notification interface in the types.tsx file to include the action property. The action property is an object of type NotificationAction, which consists of the actionType (accept, decline, view, or null) and actionUrl (a string representing the URL for the action). This change enhances the notification functionality by providing a way to specify different types of actions for each notification.