feat: Update TeamState interface to include trackQuery
The code changes in this commit update the TeamState interface to include a new property called trackQuery. This property is an object that contains information about the last document and the status of the filter match. This update allows for better tracking and management of the query in the TeamState.